Update normalize.R
Change param log to log2?
How are internal standards picked? Are they corrected per class or by mean ITSD abundance?
Is my description correct that the average ITSD is used if no corresponding ITSD is found.

Maybe we should change itsd to istd, which seems to be more commonly used.